Split up tables
