+++ src/hapi/pipelines/database/dborgtype.py
@@ -1,0 +1,12 @@
+"""OrgType table."""
+
+from sqlalchemy import Column, String
+
+from hapi.pipelines.database.base import Base
+
+
+class OrgType(Base):
+    __tablename__ = "OrgType"
+
+    code = Column(String(32), primary_key=True)
+    description = Column(String(512), nullable=False)
